pyproject.toml
[tool.poetry]
name = "secrets-service-api"
version = "0.1.0"
description = ""
authors = ["neekrasov <neekrasov@gmail.com>"]
readme = "README.md"
packages = [{include = "secrets_service_api"}]

[tool.poetry.dependencies]
python = "3.11"
aiohttp = {extras = ["speedups"], version = "^3.8.3"}
asyncpg = "^0.26.0"
aiosql = "^6.5"
python-dotenv = "^0.21.0"
passlib = "^1.7.4"
aiohttp-apispec = "^2.2.3"
dependency-injector = "^4.40.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
